+++ src/Microsoft.Extensions.Configuration/MemoryConfigurationProvider.cs
@@ -1,0 +1,47 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Collections;
+using System.Collections.Generic;
+
+namespace Microsoft.Extensions.Configuration.Memory
+{
+    public class MemoryConfigurationProvider : ConfigurationProvider, IEnumerable<KeyValuePair<string, string>>
+    {
+        private readonly MemoryConfigurationSource _source;
+
+        public MemoryConfigurationProvider(MemoryConfigurationSource source)
+        {
+            if (source == null)
+            {
+                throw new ArgumentNullException(nameof(source));
+            }
+
+            _source = source;
+
+            if (_source.InitialData != null)
+            {
+                foreach (var pair in _source.InitialData)
+                {
+                    Data.Add(pair.Key, pair.Value);
+                }
+            }
+        }
+
+        public void Add(string key, string value)
+        {
+            Data.Add(key, value);
+        }
+
+        public IEnumerator<KeyValuePair<string, string>> GetEnumerator()
+        {
+            return Data.GetEnumerator();
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            return GetEnumerator();
+        }
+    }
+}
